requirements.txt
git+git://github.com/rapptz/discord.py@master
python-dotenv
requests